backend/requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0

# Data Sources
tushare>=1.2.89
akshare>=1.12.0
requests>=2.31.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
ta

# AI Engine
google-genai>=1.0.0

# Database
SQLAlchemy>=2.0.0

# Data Validation
pydantic>=2.0.0

# Task Scheduler
APScheduler>=3.10.0

# Config
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Packaging (dev only)
pyinstaller>=6.0.0
